[libclang] Make CXUnsavedFile::Length a fixed-width uint64_t
CXUnsavedFile::Length was `unsigned long`, whose width depends on the data
model: 4 bytes under LLP64 (Windows) and 8 bytes under LP64. The struct
layout therefore differs across platforms, which corrupts the length or
crashes FFI consumers that bind the C API with a fixed-size assumption
(e.g. the OpenJDK CODETOOLS-7904079 downstream bug).

Use uint64_t so the field is consistently 8 bytes. The two internal debug
prints are switched from %ld to %llu to match. This is an intentional ABI
change to libclang's C API, documented in the release notes.

Fixes #160729.

+++ clang/tools/libclang/CIndex.cpp
@@ -4445,8 +4445,8 @@ enum CXErrorCode clang_parseTranslationUnit2FullArgv(
     for (unsigned i = 0; i != num_unsaved_files; ++i) {
       if (i)
         fprintf(stderr, ", ");
-      fprintf(stderr, "('%s', '...', %ld)", unsaved_files[i].Filename,
-              unsaved_files[i].Length);
+      fprintf(stderr, "('%s', '...', %llu)", unsaved_files[i].Filename,
+              (unsigned long long)unsaved_files[i].Length);
     }
     fprintf(stderr, "],\n");
     fprintf(stderr, "  'options' : %d,\n", options);
